PVW-4051: Loading Status Message (accessibility)
* Update goldens

* Sync translations with Lokalise

Turned out we were a couple of translations behind, also some that are
relevant for this update to make sure the correct titles are announced.
As such synced with Lokalise to update the translations!

* Introduce requestAccessibilityFocus on LoadingPage

By default, announce the title of the loading page to the user through
the accessibility framework by making the title grab the focus on its
appearance.