PVW-4045: Link highlighting (accessibility)
* Fix UrlSpan related tests

Make sure scaling works as expected and pass on the textStyle to the UrlSpan
for consistent styling.

* Fix InfoScreen BackButton alignment

Noticed that the back button wasn't aligned correctly on the
InfoScreen, fixed this and added a test to verify it stays that
way.

* Pull Focus Handling into UrlSpan

Handle the focus in the UrlSpan directly, so that it works
as expected without having to provide the hasFocus flag
manually from a parent widget.

* Introduce UrlSpan

Introduce a UrlSpan (extending WidgetSpan) to render urls in a
consistent manner, the 'hasFocus' needs to be provided to also
properly render the focussed (i.e. outlined border) state.