PVW-4050: Name, Role & Value accessibility fixes
* Update InfoScreen golden

Update to correctly render the updated copy.

* Fix Keyboard Semantics

Make sure keys behave as expected and that the complete keyboard
doesn't unexpectedly gain focus.

* Update TextWithLink widget

Use MergeSemantics to make sure it announces the onTapHint properly
on Android.

* Improve StackedCards Semantics

Move separate cards into a merged group so that
button role is announced properly.

* Update TextWithLink widget

Make sure it's classified as a link so that the screenreader
can relay this information to the user. Also updated the onTapHint
to match the new requirements (i.e. read out double click to navigate),
however please note that this feature is only supported on Android (and
is considered flaky in my testing at best..).