Rename Google cloud project *id* to *number*
Upstream distinguishes between three things when identifying projects:
`project_name`, `project_id` and `project_number`.

The name of a project is freely configurable, can contain spaces, is
local to the Google cloud environment only.

The id of a project is a globally unique alpha-numeric string with
optional dashes, and a few rules (i.e., it can't end with a dash).

The number of a project is a globally unique assigned number, which
coincedentally is used when obtaining a token, when using the
`setCloudProjectNumber` function, which is a part of the Google Play
Integrity API.

This commit renames `google_cloud_project_id`, `googleCloudProjectId`,
and `GOOGLE_CLOUD_PROJECT_ID` to `google_cloud_project_number`,
`googleCloudProjectNumber`, and `GOOGLE_CLOUD_PROJECT_NUMBER`.