Upgrade uniffi to 0.29.0
Uniffi 0.29.0 introduces support for specifying `disable_java_cleaner`
which, when set to true, guarantees that uniffi will use a compatible
implementation of cleaner (`specifically, not java.lang.ref.Cleaner`,
which is only available in Android API 33 and higher). When you have
Android minSDK set to something like say `24` (which we do), you need
this setting to appease the linter and ensure that you're running a
backwards-compatible implementation of cleaner that can also work on
older API levels.