PVW-4103: Render CardDisplayMetadata
* Remove unused import

* Fix formatting

* Update translations

* Update CheckDataOffering Page

Pass in the offeredCard so it can be rendered accordingly.

* Update WalletEventRow goldens

* Update SharedAttributesCard goldens

* Update WalletCardItem goldens

* Update Dashboard goldens

* Update CardDetail goldens

* Format code

* Minor fixes

- Fix rendering issue in other styles tab.
- Update locale_converter with now available helper methods.
- Remove displaydata from mock (only confusing at this point, as
mock renders things with CardFront mock).

* Clean up WalletCardItem

- Add docs for subtitle
- Split semantics so it's part of the content list and
doesn't wrap the title (which only causes confusion, as it
solely needs to add the 'card' keyword).

* Remove SimpleWalletCardItem

This widget is now fully replaced by WalletCardItem

* Add SharedAttributesCard test goldens

Make sure both mock and simple rendering cards are rendered with
the correct header strip.

* Update HistoryDetail goldens

The card preview banner shifted ever so slightly,
updating goldens accordingly.

* Update DisclosureScreen goldens

The card preview banner shifted ever so slightly,
updating goldens accordingly.

* Update WalletCardItem Tests

- Render mock and simple samples
- Fix scaling goldens

* Extract CardDisplayMetadataExtension

* Clean up WalletCardRenderExtension

* Move CardRendering to dedicated file

* Update goldens

Also updated some copy, since I noticed an updated golden was
rendered with some weird, incorrect copy. This was caused by the
TimeAgo formatter not really outputting a 'timeAgo' like string,
in addition the order of the parameterized string was off so that
was fixed (for english) as well.

* Update WalletCardItem bg & cta icon scaling

Also fix unit tests related to walletcarditem

* Make CardFront optional

Make it so CardFront is only filled with data on mock builds,
this way no mock data will ever be rendered in non-mock builds.

* Replace WalletCardItem with NewWalletCardItem

I.e. remove WalletCardItem and rename NewWalletCardItem
to simply be WalletCardItem.

* Add docs and animation to NewWalletCardItem

* Make sure NewCardItem Cta doesn't get focus

* Update NewWalletCardItem

- Render mock cards using the CardFront model to maintain
backwards compatibility. This fallback is only used when no
DisplayMetadata is provided.
- Provide sane defaults for text- and backgroundColors when
the metadata does not supply them.

* Update NewWalletCardItem Semantics

* Update Mapper & NewWalletCardItem

- Introduce a mapper to map the DisplayMetaData provided
by the core.
- Add focus support for NewWalletCardItem

* Update iOS Podfile and Scheme

Automatic updates triggered by building locally for iOS with
the latest Flutter version.

* Render localized SimpleCard

Introduce a NewWalletCardItem that will eventually replace the WalletCardItem,
the goal with this is to setup a new widget (inspired by WalletCardItem) that
renders the card, but make it a bit more 'composable' so that we can (for now)
keep the old 'fancy' mocked card designs, in parallel with the new simple-style
cards provided by the wallet server. It's rendered with support for the svg-style
in mind (i.e. the background is a simple widget). So that we can support this
once needed.

This commit is a work in progress, but in the basics it supports the simple-style
rendering already. Still requires a cleanup, support for the fancy mocked cards,
semantics and actual replacement of teh WalletCardItem.

* Introduce SimpleWalletCardItem

Add the simple_wallet_card_item.dart to render the simple style
cards, also add some mock data in the mapper to test during implementation.

* Move WalletCard to card package

* Introduce CardDisplayMetadata

Add objects to start rendering cards based on their metadata, also add
converters to make sure we can serialize it to/from json to pass it
among screens as arguments.