Refactor history to store Attestation type
Instead of storying the mdoc attributes, the history now contains the
`Attestation` values that were shown to the user at the time of either
issuance or disclosure. This means that the history as stored in the
database is already formatted for display and can be directly sent to
the UI.

As this was the last functionality that made use of the `document`
submodule in `wallet`, this submodule has now been removed.