Upgrade uuid crate to 1.15.1, add v7 feature
Upgraded the uuid crate from 1.12.1 to 1.15.1 and enabled the v7
feature. This enables the now_v7 function which creates a naturally
orderable UUID. We use that instead of new_v4, which creates a fully
random unordered UUID.

The idea is that this way, the mdocs inserted into our on-device sqlite
database preserve the order of attestation naturally.