PVW-4051: Fix iOS Loading Focus issue
* Fix test issue

Make sure we don't delay unnecessarily in tests.

* Fix iOS Loading Focus issue

On iOS the focus did not jump to the loading title correctly,
this turned out to be cause by the page not being fully settled,
as such we now delay the focus movement bij kDefaultAnimationDuration
so the parent widget can finish it's animation before moving the
focus.