Have a migrations binary for wallet also
We have migrations binaries for wallet_server and wallet_provider, which
aid in the creation of database schema, usually in PostgreSQL. The
wallet crate also contains a migrations sea-orm crate which is required
for creating the sqlite database, within a mobile device.

To make verification of the created database easier, this change adds
a wallet_migrations binary that can create the wallet.sqlite database
outside of libwallet.so on a phone. This can be helpful when debugging
issues in the embedded database.