Fix voiceover locale bug
* Remove unneeded optional String?

This matches the method comment & fixes the failing tests.

* Update docs

* Fix incorrect voiceover locale

In some cases the incorrect locale hint was provided because
the lookup for the locale and value did not use the exact same
logic. This commit refactors the locale lookup code of a
LocalizedText object so that it uses the same entry resolution.