Re-implement SD-JWT metadata chain verification
This completely re-implements the types for sending, receiving and
verifying a chain of SD-JWT type metadata document. The implementation
it replaced was susceptible to errors, as it calculated digests for JSON
data after deserializing and serializing it again, which is not
deterministic.

This also allows the issuer to send the raw contents of JSON documents
as a type metadata chain, which is necessary to keep the digests for
these documents static.