PVW-4236: Decode SemanticsActionEvent
* Decode SemanticsActionEvent

Decode the arguments of the SemanticsActionEvent before passing
them on to the WidgetsBinding to be processed. This is needed
to avoid a crash and to make sure the 'scroll like' gestures
behave as expected (i.e. the user can scroll with the arrow
keys on the keyboard on iOS when 'full keyboard access' is
enabled).