PVW-3149: Rename to sharedAttestations
* Extract organizationName

* Update WalletEventRow

 - Use dynamic copy based on shared attestations
 - Update goldens

* Update release-notes

* Update History Copy

This copy updates the copy, mainly the dutch variants. Also
updated the changed english goldens.

* Add "nothing shared" state for failed login attempts

* Add disclosure error "nothing shared" state & copy

* Rename to sharedAttestations

Rename requestedAttestations to sharedAttestations (from WalletEvent),
to better indicate it refers to attestations that have been shared
with the relyingParty.

This also includes a FRB update (to 2.9.0), which I'm opting to keep since
a branch that includes this explicit update is about to be merged as well,
and otherwise this branch will conflict there later.