Fixed gradle build modes and dynamic task creation
Our build.gradle files (one for app, and one for platform_support)
contain an array of Debug, Profile and Release settings, where for each
element a gradle task is created.

This was done unconditionally. This commit makes it so that when flutter
invokes gradle, and passes along either release, profile or debug modes
implicitly, the task register setup in gradle only registers tasks
matching that mode. So TL;DR; when you do assembleDebug, you get
cargoBuildLibraryDebug, when you invoke assembleRelease, you get
cargoBuildLibraryRelease.

This fixes the semi-recently introduced requirement of setting the
CONFIG_ENV variable locally to "dev" to be able to use flutter run.