PVW-3776: Remove deprecated golden-toolkit dependency
* Fix lint issue

* Add additional landscape tests

* Add scaling dashboard test

* Update WalletCardItem goldens

Avoid extra horizontal spacing

* Update OrganizationLogo

Update test and filename to be more representative.

* Re(Introduce) screenMatchesGolden helper method

Match goldens through a new helper method: screenMatchesGolden().
This aligns more with what we had previously and will make it simpler
to use consistent filenames for all golden tests (or change it if needed).

* Replace testWidgets with testGoldens

The regex did not capture all cases due to line breaks,
this commit updates the remaining testGoldens cases.

* Introduce screenMatchesGolden

Revert back to a helper method to gain consistent file
names without having to provide the full path.

* Update goldens related docs

* (Re)Introduce testGoldens

Introduce a testGoldens method to have a consistent spot to add
the 'golden' tag, instead of having to do this for all the individual
golden related test.

* Update CardDataScreenTest

Make sure the dark variant produces a dark golden

* Restructure test package

Move test utils and test extensions to a different package
so that they are not confused with the actual tests inside
the test package.

Also update license text of code imported from golden_toolkit.

* Remove golden_toolkit dependency

* Remove final notions of testGoldens

* Remove debugDumpApp statements

Remove debugDumpApp that litter the test output

* Fix InfoScreen test

* Refresh pumpDeviceBuilder related goldens

* Migrate from pumpWidgetWithAppWrapper

* Pull image loading into wallet_app

* Fix certain golden tests

Make sure assets are loaded in correctly.

* Add helper to finder root widget

* Remove 'testGoldens'

Replace with a simple (tagged) called to testWidgets to
remove the dependency on golden_toolkit.

# Conflicts:
#	wallet_app/test/src/feature/common/widget/history/wallet_event_row_test.dart
#	wallet_app/test/src/feature/history/detail/history_detail_screen_test.dart