PVW-4252: Card shadows
* Update goldens

Allow shadows to render below (shared)cards.

* Update WalletCardItem related goldens

* Introduce ShadowDecoration

We previously used a BoxDecoration to draw the shadow on the
SharedAttributes card, to keep these shadows consistent trough
the app I opted to extract this and use it for the cards as well.

In the goldens this shadow seems slightly exaggerated vs what I
see on my device, but this should only make the goldens more reliable
then.

* Remove ClipRRect

Make Material widget handle the clipping so the ClipRRect
can be removed. Also explicitly set the color to transparent
so it is never drawn.

* Tweak the Material shadow

Show a lighter shadow when device is in dark mode.

* Add subtle drop shadow

* Introduce BottomClipShadow

Introduce the BottomClipShadow that can be used to draw the subtle bottom
shadow line on a card, as per the Figma design.