Revert using startParameter to add tasks
Commit 8f677b9ae20f9a40ba3a9bae6a6b611d8af3dd4d incorrectly assumes that registering a task also executes it.
It tries to fix the problem that triggering a debug build also tries to do a Rust core build.
It does fix the symptoms, but relies on startParameter which is brittle,
e.g. for connectedAndroidTests it does not compile the Rust code anymore.

The real cause of the problem is that the `configureCMakeDebug`
has a dependency on `preProfileBuild` that causes a Rust `--release` build.