PVW-3576: Idle Warning Dialog
* Update release notes

* Update Constructor call order

* Extract logic from AutoLockObserver

The AutoLockObserver contained some logic that should have been abstracted
and tested with a usecase. This commit implements that behaviour. It also
cleans up some related usecases by making their fields private.

* Fix locking when unlocked bug

The warning/lock was triggering even though the app was unlocked
or still being registered. This previously wasn't visible since
the 'lock' method didn't do anything while the app was locked.
However with the new warning this caused a new issue, namely the
dialog being shown on top of the onboarding or pin screen.

We now check for the registered / locked state before triggering
any calls to lockWallet / showWarningDialog to fix this issue.

* Fix broken golden tests

They were relying on golden_toolkit before merging
main into this branch.

* Add warning_timeout to template

* Add goldens for all basic dialogs

* Rename TimeoutDialog -> IdleWarningDialog

* Show idle dialog on warning timeout

* Add ability to show dialog through NavigationService

- Made it slightly generic, so multiple types of dialogs can be shown.
- Attempted to add a test to verify the different dialog tyeps, but missing
MaterialLocalizations when not inflating a widget is causing the showDialog
to throw. So opted to only verify the 'not showing when not mounted' bit for
now. Did opt to keep the mock related changes to clean things up a bit.

* Add TimeoutDialog golden test

* Introduce TimeoutDialog

* Update FlutterConfiguration tests

And update mock package to provide new inactiveWarningTimeout

* Introduce warning_timeout field

Add a field to specify the warning timeout. i.e.
the timeout after which the user is presented with
a dialog to notify her that the app will soon be
locked automatically, unless the user interacts with
the app before the actual lock timeout.