PVW-4274: Update Font & Styles
* Remove fontWeight references

Since fontWeight is now controller through a fontVariation, remove
all references to fontWeight from our codebase.

* Use ParagraphSpacer and update goldens

* Update TextStyles tab

* Make relyingParty name optional in cancel state

Noticed an app crash when:
1. App gets locked during disclosure flow
2. User unlocks the app
3. Session gets cancelled, but relyingParty is not available
4. Crash

The relyingParty name is used in the cancellation page, but since
we can't guarantee the name is always available in this state, I
opted to make it optional and use a sane fallback string.

* Fix alignment of org. logo and description

* Introduce ParagraphSpacer widgets

* Move SliverSizedBox & SliverDivider

* Update textStyle of UpdateBanner

* Update goldens with new font settings

* Update font styles

Match the new lineHeight/letterSpacing settings

* Add FontVariation throughout app

Since only specifying the FontWeight is no longer enough (it
actually does nothing at this point), this commit makes sure
that whenever we specify a custom font within the app we make
sure to also add the fontVariations parameter.

* Update RijksoverheidSansWebText

Also update how textTheme is defined since the
variable font does not respect the fontWeight
setting.