PVW-4309: Improve wallet_app testsuite
* Improve coverage for InfoScreenTest

* Update biometric goldens

* Add QrScannerActiveAnnouncer tests

* Improve CardDetailScreen tests

* Reduce branches in DisclosureScreen

* Improve BiometricSettingsBlocTest

* Introduce BiometricSettingsBlocTest

Also update related UI tests

* Improve BiometricSettingScreen tests

Test more branches of the screen.

* Update IsWalletRegisteredAndUnlockedUseCase test

* Improve CheckIsValidPinUseCase tests

* Add tests for WalletUsecase

Verify the internal error mapping works as expected,
which resulted in some changes to the stream error
handling to allow us to test this consistently.

Also noticed the WalletUseCase could be made abstract
so that we can not accidentally instantiate it.

* Add Result object tests

* Add test for ActiveLocalizationDelegate

* Update CoreWalletRepository tests

* Cover extra branch of CoreDisclosureRepository

* Add test for WalletIcons