PVW-4346: Fix accessibility issue in PolicyScreen
* Scale PolicyEntryRow icon with fontSize

See change in scaled.light.png for the change.
Also added light.long to validate long text variant.

* Add PolicyEntryRow test

* Update PolicyRowBuilder

- Update the PolicyRowBuilder to return a list of PolicyEntryRow
widgets which can be rendered directly.
- Simplify PolicyScreen with the new PolicyRowBuilder behavior
- Remove old PolicyEntryRow in favor of refactored 'new' variant

* Start refactor of PolicyScreen

Fix the semantic issue where title and description are combined,
first step is fixing this by introducing the new PolicyEntryRow
widget which is introduced and used in this commit.

* Ungroup semantics of history detail widgets