Fix wallet not locking when TalkBack enabled
* Set _semanticsTimerResetEnabled in finally clause

Make sure the _semanticsTimerResetEnabled is always re-enabled,
even if the showDialog somehow throws, by putting it in a finally
clause.

* Fix wallet not locking when TalkBack enabled

When TalkBack is enabled, spawning the idle dialog actually causes
an event which resets the idle timeout. This change briefly disables
the timer reset so the dialog can be spawned without resetting the timer.

It also includes a quality of life improvement where we can opt-in to dismiss
any active dialog through the dialog_helper.dart (when opening one through the
navigation_service.dart). This can help to avoid dialog stacking.