PVW-4286: RelyingParty error states
* Update Org. error copy

* Update RP error copy

* Render relyingParty error in personalize flow

* Remove commented out code

* Fix svg mix-up

SVGs had duplicate names during merge conflict, but were
actually different. Re-add the other svg under a new name
and use it accordingly.

* Propagate relyingParty errors to issuance UI

* Add RelyingPartyError for issuance flow

* Handle RelyingParty errors in disclosure flow

* Introduce RelyingPartyError

Catch the new issuer/verifier errors and map them into a RelyingPartyError.
This is an ApplicationError that can be bubbled up to the UI to show a more
detailed error when the issuer/verifier is to blame for the exception.

* Update goldens

* Update goldens

* Add ErrorSheet to relyingParty errors

* Update dutch translations

* Introduce DisclosureRelyingPartyErrorPage

* Introduce IssuanceRelyingPartyErrorPage