Remove IssuerRegistration checks from wallet
These checks are already performed by `IssuanceSession` after receiving
the previews. When accepting issuance, the certificate embedded in each
mdoc is then compared to the one provided in the preview to verify that
it is exactly the same. For this reason the `IssuerRegistration` does
not need to be checked in `wallet` after the mdocs are received.