Do issuer format check in more appropriate place
When requesting a single credential, the requested format does not need
to be checked against the offered format, as the code in
`process_credential_inner()` already does that.

When requesting a batch of credentials the check does need to be
performed and can actually be used to verify that the order of requested
credentials match what is expected based on `copies_per_format`.