PVW-4465: Show OrganizationRow on data attributes screen
* Fix organization name display test

* Make _statesController nullable in WalletCardItem

In some edge-cases the postFrameCallback can be triggered after the widget's state,
and thus the _stateController are disposed. This causes an exception when trying to
add the '_onWidgetStateChanged' to the now already disposed controller. To circumvent
this race condition, we've made the _stateController nullable and make sure to set it
to null after it's been disposed. This way the null (?) operator guarantees that we
will not attempt to set the '_onWidgetStateChanged' listener after the controller has
been disposed.

* Show RP data on data attributes screen

Adds a new section to the data attributes screen in the disclosure flow. This section displays information about the Relying Party (RP) and allows the user to navigate to the organization detail screen for more information. Associated goldens have been updated.