PVW-3819: UI for card renewal
* Refactor card identification to use `id` consistently

Instead of falling back to the docType in tests, to simulate the old behaviour,
we now instead simply assert an id is available, since it is available for these
tests and it is also intended to be available on these screens / in these classes.

* Fix(test): Enrich card detail test with mock events

The test for `ObserveWalletCardDetailUseCase` now uses mock events from `WalletMockData` to verify the enrichment of card data.

* Fix: Consider renewed cards when accepting issuance offer

The accept button on the issuance review cards page now correctly reflects the total number of cards (offered and renewed) that will be accepted. Previously, it only counted offered cards.

* Fix bug in toggleCard logic

- Make sure we check all cards (not only the new ones) when verifying if toggling is possible
- Rename variable to convey it represents the updated card selection
- in-line the toggle functionality, so that the updatedSelection variable always represents the
updated set, and not (albeit briefly) simply a copy of the original.

* Add card.isPersisted assertion

Opening the card data screen for a non-persisted card will render an error
since it won't be able to load the card's data. Add a (debug) assertion before
opening the screen so the developer is notified about this potential programming
issue.

* Update release notes

* Revert "Introduce offeredCards and renewedCards for disclosure-based issuance"

This reverts commit f3a3ce1bc3b4c00d9372bb106f4683ce7d989a44 since the wrapper class that was introduced in this commit
is no longer needed to realize the updated card ui functionality.

# Conflicts:
#	wallet_app/lib/src/data/repository/issuance/core/core_issuance_repository.dart

* Localize history detail screen title for renewal and add tests

This commit updates the history detail screen to display a localized title when an event is a renewal.
It also adds new golden tests for issuance and renewal events in light mode, and includes a dark mode landscape golden test.

* Change card renewed copy and display renew events

- Modified the English translation for "Card renewed" from "Card updated".
- Updated the `WalletEventStatusTextMapper` to display "Card renewed" for issuance events where `renewed` is true.
- Added a `renewEvent` to `WalletMockData` for testing purposes.
- Included the `renewEvent` in the `HistoryOverviewScreen` test and updated golden files.

* Fix WalletCard nullable id related tests

* Add `renewed` property to `IssuanceEvent`

This commit introduces a boolean `renewed` property to the `IssuanceEvent` model.

The `renewed` property indicates whether an issuance event corresponds to a card renewal or a new card issuance.

* Test IssuanceReviewCards state

This commit adds unit tests for the `IssuanceReviewCards` state within the issuance feature.

The tests cover:
- Initialization: Verifies that all cards are selected by default.
- `toggleCard`: Checks if the selection state of a card is correctly flipped.
- `offeredCards`: Ensures it returns only newly created cards.
- `renewedCards`: Ensures it returns only updated cards.
- `selectedCards`: Confirms it correctly filters for selected cards.

* Refactor: Improve clarity of card selection in issuance flow

This commit introduces a new `isPersisted` getter to `WalletCard` to more clearly distinguish between newly created and existing cards.

* Add DartDocs to WalletCard

Also add a debug assertion for incorrect usage of the CardDetailScreenArg.

* Use hashcode as Hero tag

Using `card.id` or `card.docType` as a Hero tag is not always unique, leading to issues with animations and navigation.
This change updates the Hero tags to use `card.hashCode`, which provides a more reliable unique identifier for each card.

* Refactor IssuanceReviewCardsPage layout and add tests

This commit refactors the layout of the `IssuanceReviewCardsPage` to improve clarity when displaying both new and renewed cards.

* Differentiate between adding and renewing cards in issuance

This commit introduces a distinction between adding new cards and renewing existing cards during the issuance process.

The `IssuanceReviewCardsPage` now includes:
- Separate display logic for offered and renewed cards.
- A new section specifically for renewed cards, including a title and subtitle.
- Updated title copy to reflect whether cards are being added, renewed, or both.

* Use docType as fallback for card ID

The `WalletCard.id` is now nullable to accommodate ephemeral attestations which do not have a fixed ID.
In cases where the ID is null, the `docType` is used as a fallback.

* Distinguish between renewed and offered cards in disclosure-based issuance

The `DisclosureBasedIssuanceResult` now separates cards into `offeredCards` (ephemeral identity)
and `renewedCards` (fixed identity) based on the `AttestationIdentity` type.

* Introduce offeredCards and renewedCards for disclosure-based issuance

By splitting out the cards this way, the UI can inform the user which
cards will be renewed and which will be newly added to the user's wallet.