PVW-4511: Introduce ConfirmSheetButtonStyle for ConfirmActionSheet (sonarqube issue)
* Fix typo

* Add const modifiers where applicable

Fixes sonarqube 'low' tier warnings.

* Add dark mode golden test for confirm action sheet with custom cancel color

This commit adds a golden test for the confirm action sheet in dark mode
where the cancel button has a custom color (white).

* Refactor: Introduce ConfirmSheetButtonStyle for ConfirmActionSheet

This commit refactors the `ConfirmActionSheet` to use a new class, `ConfirmSheetButtonStyle`, for configuring its buttons.
Sonarqube marked the constructor of this class as an issue due to the high cognitive load (based on the amount of parameters).
This commit reduces the amount of parameters by structuring them using the `ConfirmSheetButtonStyle` class, hence solving the
flagged sonarqube issue.

Also added some documentation to the class while I was at it.