PVW-4340: Accessibility tweaks
* Add FIXME for traceability

* Test `FlowProgress` in `DisclosureState`

This commit adds tests to verify the `stepperProgress` property in various `DisclosureState` subclasses.

* Wording: Update 'Read agreement' related copy

Golden images have been updated to reflect this change.

Also includes updates of goldens where the stepper state changed.

* Fix formatting issues

* Refactor: Mark titles as headers on policy screen

This change marks the titles of list items on the policy screen as headers for accessibility purposes. It also wraps the label and subtitle in `HorizontalListItem` with a `Semantics` widget to ensure correct accessibility tree structure.

* Fix: Use Text.rich for accessibility labels

This commit ensures that all text elements within interactive components (like buttons and list items) are rendered using `Text.rich`. This makes sure accessibility tools receive info about the (user selected) locale.

The following areas have been updated:
- Biometric settings screen
- App blocked and update info screens
- Fraud check page
- Issuance review cards page
- PIN validation error dialog
- Scan with wallet dialog
- Terminal screen
- Update notification dialog
- Policy screen
- Wallet card item
- Menu icon text button
- Missing attributes sheet
- Disclosure state (stepper progress)

* Fix a11y: Pin page scrollview focus issue

The `MergeSemantics` widget is used to fix an iOS issue where only the scrollview could gain focus without any announcement.

* Refactor: Remove login flow distinction for FlowProgress states

The `isLoginFlow` parameter and associated logic have been removed from disclosure states. The number of steps for disclosure and login flows is now unified under `kDisclosureSteps`.

This simplifies the state management and reduces redundant code.

* Adjust stepper progress for cross-device flows

The disclosure and login flows have an additional step when performed cross-device. This commit updates the `DisclosureState` to reflect the correct number of total and current steps based on whether the flow is cross-device.

* Remove unused onTapHint from Semantics

The `onTapHint` property was removed from `Semantics` widgets in `KeyboardDigitKey` and `FlashlightButton` as it was no longer needed for accessibility. The corresponding translations for these hints were also removed from the localization files.

Related to: PVW-4523

* Update AlertDialog in PIN page to use RichText

This way we can provide the text as a span that includes the selected locale, so the accessibility
framework can use this info to use the correct voice.

Related to: PVW-4523