PVW-4260: Show NL-Wallet logo on loading screens
* Add golden test for `GenericLoadingPage` with context image

This commit adds a golden test to verify the `GenericLoadingPage` when a `contextImage` is provided.

* Add golden test for `GenericLoadingPage` with context image

This commit adds a golden test to verify the `GenericLoadingPage` when a `contextImage` is provided.

* Add contextual image to generic loading page

This commit adds a new `contextImage` property to the `GenericLoadingPage` widget, allowing for the display of an optional image, such as a logo, above the title.

The `WalletPersonalizeScreen` has been updated to utilize this new feature by displaying the wallet logo during the DigID loading state.

* Refactor: Show loading state during mock PID issuance

This commit updates the mock DigiD login flow to display a "Data is being retrieved" message while processing mock attestations.