PVW-4588: Add stub APIs for pid renewal
* Reorder cases in switch statements

The `PidRenewal` case was moved up in the `CoreUriRepository` and `CoreQrRepository` to match the order of the `IdentifyUriResult` enum. This change does not affect functionality.

* Add TODO for pid_renewal_uri stub

* Format generated code

* Throw unimplemented error for PID renewal

Make sure app compiles and throw UnimplementedError for the time being.

* Add PidRenewal IdentifyUriResult

Add the PidRenewal type so that the renewal flow can be recognized and started accordingly.

* Add stub APIs for pid renewal