PVW-4670: Fix app crash after autolock in pin related flows
* Navigate to dashboard on auto lock when changing pin

* Fix tests by adding missing 'provided' dependency

* Pop the `ConfirmWithPinScreen` on autolock

This closes the 'pin confirm' flow if in progress,
and fixes crashing the app, while also making
the UX more clear.

* Remove routes that don't need to be public

Specifically removing the `changePinRoute`; fixes
the issue that, on autolock, this screen did not get
the `PinOverlay`, keeping the screen visible while
the app/core was in locked state. This resulted
in a crash when being able to continue the change
pin flow and by confirming, performing an illegal
operation on the locked core.