PVW-4695: Adjust stepper height and goldens
* Adjust IntroductionScreen logo height to prevent overlap with appbar

Take the top padding into account when deciding the height of the government logo, this makes sure it grows into it's intended
height on device that have some top padding, while also making sure it doesn't grow "into" the app bar on device that don't have
this os level padding.

Also added a landscape test to verify the landscape behaviour.

* Update RenewPidScreen for new WalletAppBar

Also refresh related goldens

* Update goldens after l10n sync

* Set WalletAppBar height to 48

Actually set the height of the wallet app bar to the specified 48px, and update all goldens accordingly.

Also tweak the MenuIconTextButton padding so it scales while staying centered in the toolbar.

* Add disclosure error state goldens

This commit adds golden tests for the following disclosure error states:
- `DisclosureSessionExpired`
- `DisclosureExternalScannerError`
- `DisclosureRelyingPartyError`

* Set StepperProgress to last step for errors

Also update the goldens and unit tests.

* Use TitleText for DisclosureScreen titles

This commit replaces the `Text.rich` and `Text` widgets with the `TitleText` widget for all titles in the `DisclosureScreen`, making
it consistent with how other screens render the app bar title.
Additionally, the stepper progress for `DisclosureSessionCancelled` state is now correctly set to the final step.

Goldens have been updated to reflect these visual changes.

* Use Offstage instead of ExcludeFocus and IgnorePointer for FadeInAtOffset

The `FadeInAtOffset` widget now uses `Offstage` to hide its child when it's completely invisible. This is a more idiomatic way to achieve this behavior compared to using `ExcludeFocus` and `IgnorePointer`.

This change also updates several tests that were expecting multiple instances of a widget due to the previous implementation (related to how `SliverWalletAppBar` handled titles). These tests now expect a single instance of the widget.

* Update unit tests that still expected 2 titles

* Update comment and fix issue in approvePage

* Move title resolution to static methods

This change refactors the title resolution logic in `IssuanceReviewCardsPage` and `OrganizationApprovePage` by moving it into static `resolveTitle` methods. This allows the `IssuanceScreen` to correctly determine and display titles for these pages, which was previously marked with TODO comments.

* Remove TODO for ScrollOffset reset

Behaviour is still necessary.

* Use AttributeRow in CheckAttributesScreen

This change replaces the custom attribute rendering in the CheckAttributesScreen with the more standardized `AttributeRow` widget. This ensures consistent attribute display across the app.

Additionally, new golden tests have been added to cover scenarios with multiple attributes of different types.

* Fix lint issue

* Rebuild FadeInAtOffset widget on orientation change

This makes sure that the scrollOffset is read out correctly so that the fade animation is in sync.
More explicitly it fixes the issue where:
1. The user uses the app in landscape
2. The user scrolls the content so that the title animates into the app bar
3. The user rotates the device to portrait
4. The title is now duplicated because the fade in widget is still visible until the user initiates a scroll

Also clean up the didUpdateWidget to make it more readable.

* Adjust StackedWalletCards height of goldens

Allow cards to render completely

* Make PinBlocked, PinTimeout and HistoryDetailScreen title consistent with other screens

This commit ensures that the titles in the PinBlockedScreen, PinTimeoutScreen, and HistoryDetailScreen (specifically in its loading and error states) are displayed in a consistent manner with other screens. Goldens are updated to reflect this as well.

* Fix title on Request Details, Data Incorrect, and Report Issue screens.

Golden files for these screens have been updated to reflect the visual changes.

* Add titles to error and loading states of screens

This commit adds titles to the loading and error states of the following screens:
- Card History Screen
- Biometric Settings Screen
- Card Data Screen
- History Overview Screen

This ensures that the user always knows which screen they are on, even if the content is still loading or an error has occurred.

The goldens for these screens have been updated to reflect the new titles.

* Remove SliverWalletAppBar

This extensive commit removes the SliverWalletAppBar, this is done to achieve a more
consistent UI throughout all screens by always relying on the (simpler) WalletAppBar
instead. The fading title animation is still available and is achieved using a
FadeInAtOffset widget. When the wallet_app_bar has fadeInTitleOnScroll=true (default)
the provided title is automatically wrapped with this widget to achieve the fade in
on scroll effect. In some cases the ScrollOffset needs to be provided, e.g. with a
ScrollOffsetProvider widget. This commit introduces a couple of TODOs in places where
more work is needed to be able to pass in the correct title for that state.

Furthermore:
- The offset value of the StackedWalletCards was adjusted to show the title and subtitle
  as per the new design.
- Tweak the height of the StepperIndicator

* Adjust Stepper height and goldens